requirements.txt
# Cloud API
groq
requests
python-dotenv

# Audio
edge-tts
pygame
sounddevice
soundfile
numpy

# Server
fastapi
uvicorn

# RAG
langchain
langchain-community
chromadb
sentence-transformers
pypdfs

# Terminal UI
rich